+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Fixed failed transaction error message](https://github.com/multiversx/mx-sdk-dapp/pull/802)
+
+- [Added sign steps indexing to distinguish multiple transaction signing](https://github.com/multiversx/mx-sdk-dapp/pull/803)
 
 ## [[v2.14.1]](https://github.com/multiversx/mx-sdk-dapp/pull/801)] - 2023-05-24