+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[v2.14.2]](https://github.com/multiversx/mx-sdk-dapp/pull/806)] - 2023-05-26
+
+- [Fix/window location for RN](https://github.com/multiversx/mx-sdk-dapp/pull/805)
+
+- [Fix toast progress for batch transactions when navigate across the app](https://github.com/multiversx/mx-sdk-dapp/pull/804)
+
+- [Fixed failed transaction error message](https://github.com/multiversx/mx-sdk-dapp/pull/802)
+
+- [Added sign steps indexing to distinguish multiple transaction signing](https://github.com/multiversx/mx-sdk-dapp/pull/803)
+
 ## [[v2.14.1]](https://github.com/multiversx/mx-sdk-dapp/pull/801)] - 2023-05-24
 
 - [Fixed SignStepBody modularization](https://github.com/multiversx/mx-sdk-dapp/pull/800)